deleted redundant code components, and moved all the code containers into one file
